zkp-component: Add prover feature and prune tokio features
Add a new `prover` feature for the `zkp-component` crate such that
it only requires `process` and `io-utils` features from tokio when
enabled.
Also pruned the `tokio` features to stopped using `full` since it
doesn't compile for wasm.
