+++ lib/Cargo.toml
@@ -83,3 +83,4 @@ panic = ["log-panics"]
 rpc-server = ["validator", "nimiq-rpc-server", "nimiq-wallet"]
 validator = ["nimiq-validator", "nimiq-validator-network", "nimiq-rpc-server"]
 wallet = ["nimiq-wallet"]
+zkp-prover = ["nimiq-zkp-component/prover"]
