+++ zkp-component/src/lib.rs
@@ -1,8 +1,12 @@
 pub use zkp_component::ZKPComponent;
 
+#[cfg(feature = "prover")]
+pub mod proof_gen_utils;
 pub mod proof_utils;
+#[cfg(feature = "prover")]
 pub mod prover_binary;
 pub mod types;
 pub mod zkp_component;
+#[cfg(feature = "prover")]
 pub mod zkp_prover;
 pub mod zkp_requests;
